Bunch of enemy changes

+++ ApocalypseScrapper/Assets/Scripts/TurretAI.cs
@@ -147,9 +147,9 @@ public class TurretAI : MonoBehaviour, IDamage
             playerInRange = false;
         }
     }
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
-        HP -= amount;
+        HP -= (int)amount;
 
         // if we (the enemy) gets shot, we should know where the player shot us from
         //agent.SetDestination(gameManager.instance.player.transform.position);
